Improved CreateGroup
