+++ Cumulus.Monads/Graph/CreateGroup.cs
@@ -15,6 +15,8 @@ using Microsoft.Azure.WebJobs.Host;
 using Microsoft.Graph;
 using Cumulus.Monads.Helpers;
 using Group = Microsoft.Graph.Group;
+using System.Linq;
+
 
 namespace Cumulus.Monads.Graph
 {
@@ -22,9 +24,7 @@ namespace Cumulus.Monads.Graph
     {
         private static readonly Regex ReRemoveIllegalChars = new Regex("[^a-z0-9-.]", RegexOptions.Compiled | RegexOptions.IgnoreCase);
 
-        [FunctionName("CreateGroup")]
-        [ResponseType(typeof(CreateGroupResponse))]
-        [Display(Name = "Create Office 365 Group", Description = "This action will create a new Office 365 Group")]
+        [FunctionName("CreateGroupWithOwnerMember")]
         public static async Task<HttpResponseMessage> Run([HttpTrigger(AuthorizationLevel.Function, "post")]CreateGroupRequest request, TraceWriter log)
         {
             try
@@ -40,7 +40,9 @@ namespace Cumulus.Monads.Graph
                 string mailNickName = await GetUniqueMailAlias(request);
                 string displayName = GetDisplayName(request);
                 GraphServiceClient client = ConnectADAL.GetGraphClient(GraphEndpoint.Beta);
-                var newGroup = new Group
+
+
+                var newGroup = new GroupExtended
                 {
                     DisplayName = displayName,
                     Description = GetDescription(request.Description, 1000),
@@ -52,12 +54,53 @@ namespace Cumulus.Monads.Graph
                     Classification = request.Classification
                 };
 
+
+                if (request.owners != null && request.owners.Length > 0)
+                {
+                    var users = GetUsers(client, request.owners);
+                    if (users != null)
+                    {
+                        newGroup.OwnersODataBind = users.Select(u => string.Format("https://graph.microsoft.com/v1.0/users/{0}", u.Id)).ToArray();
+                    }
+
+                }
+
+                if (request.members != null && request.members.Length > 0)
+                {
+                    var users = GetUsers(client, request.members);
+                    if (users != null)
+                    {
+                        newGroup.MembersODataBind = users.Select(u => string.Format("https://graph.microsoft.com/v1.0/users/{0}", u.Id)).ToArray();
+                    }
+                }
+
+
+
                 var addedGroup = await client.Groups.Request().AddAsync(newGroup);
+
+              
+                var groupToUpdate = await client.Groups[addedGroup.Id]
+                        .Request()
+                        .GetAsync();
+
+                if (request.members != null && request.members.Length > 0)
+                {
+                    // For each and every owner
+                    await UpdateMembers(request.members, client, groupToUpdate);
+                }
+
+                if (request.owners != null && request.owners.Length > 0)
+                {
+                    // For each and every owner
+                    await UpdateOwners(request.owners, client, groupToUpdate);
+                }
+
                 var createGroupResponse = new CreateGroupResponse
                 {
                     GroupId = addedGroup.Id,
                     DisplayName = displayName,
                     Mail = addedGroup.Mail
+
                 };
                 try
                 {
@@ -90,7 +133,200 @@ namespace Cumulus.Monads.Graph
                     Content = new ObjectContent<string>(e.Message, new JsonMediaTypeFormatter())
                 });
             }
+
         }
+
+
+
+        private static async Task UpdateMembers(string[] members, GraphServiceClient graphClient, Group targetGroup)
+        {
+            foreach (var m in members)
+            {
+                // Search for the user object
+                var memberQuery = await graphClient.Users
+                    .Request()
+                    .Filter($"userPrincipalName eq '{m}'")
+                    .GetAsync();
+
+                var member = memberQuery.FirstOrDefault();
+
+                if (member != null)
+                {
+                    try
+                    {
+                        // And if any, add it to the collection of group's owners
+                        await graphClient.Groups[targetGroup.Id].Members.References.Request().AddAsync(member);
+                    }
+                    catch (ServiceException ex)
+                    {
+                        if (ex.Error.Code == "Request_BadRequest" &&
+                            ex.Error.Message.Contains("added object references already exist"))
+                        {
+                            // Skip any already existing member
+                        }
+                        else
+                        {
+                            throw ex;
+                        }
+                    }
+                }
+            }
+
+            // Remove any leftover member
+            var fullListOfMembers = await graphClient.Groups[targetGroup.Id].Members.Request().Select("userPrincipalName, Id").GetAsync();
+            var pageExists = true;
+
+            while (pageExists)
+            {
+                foreach (var member in fullListOfMembers)
+                {
+                    var currentMemberPrincipalName = (member as Microsoft.Graph.User)?.UserPrincipalName;
+                    if (!String.IsNullOrEmpty(currentMemberPrincipalName) &&
+                        !members.Contains(currentMemberPrincipalName, StringComparer.InvariantCultureIgnoreCase))
+                    {
+                        try
+                        {
+                            // If it is not in the list of current owners, just remove it
+                            await graphClient.Groups[targetGroup.Id].Members[member.Id].Reference.Request().DeleteAsync();
+                        }
+                        catch (ServiceException ex)
+                        {
+                            if (ex.Error.Code == "Request_BadRequest")
+                            {
+                                // Skip any failing removal
+                            }
+                            else
+                            {
+                                throw ex;
+                            }
+                        }
+                    }
+                }
+
+                if (fullListOfMembers.NextPageRequest != null)
+                {
+                    fullListOfMembers = await fullListOfMembers.NextPageRequest.GetAsync();
+                }
+                else
+                {
+                    pageExists = false;
+                }
+            }
+        }
+
+
+        private static async Task UpdateOwners(string[] owners, GraphServiceClient graphClient, Group targetGroup)
+        {
+            foreach (var o in owners)
+            {
+                // Search for the user object
+                var ownerQuery = await graphClient.Users
+                    .Request()
+                    .Filter($"userPrincipalName eq '{o}'")
+                    .GetAsync();
+
+                var owner = ownerQuery.FirstOrDefault();
+
+                if (owner != null)
+                {
+                    try
+                    {
+                        // And if any, add it to the collection of group's owners
+                        await graphClient.Groups[targetGroup.Id].Owners.References.Request().AddAsync(owner);
+                    }
+                    catch (ServiceException ex)
+                    {
+                        if (ex.Error.Code == "Request_BadRequest" &&
+                            ex.Error.Message.Contains("added object references already exist"))
+                        {
+                            // Skip any already existing owner
+                        }
+                        else
+                        {
+                            throw ex;
+                        }
+                    }
+                }
+            }
+
+            // Remove any leftover owner
+            var fullListOfOwners = await graphClient.Groups[targetGroup.Id].Owners.Request().Select("userPrincipalName, Id").GetAsync();
+            var pageExists = true;
+
+            while (pageExists)
+            {
+                foreach (var owner in fullListOfOwners)
+                {
+                    var currentOwnerPrincipalName = (owner as Microsoft.Graph.User)?.UserPrincipalName;
+                    if (!String.IsNullOrEmpty(currentOwnerPrincipalName) &&
+                        !owners.Contains(currentOwnerPrincipalName, StringComparer.InvariantCultureIgnoreCase))
+                    {
+                        try
+                        {
+                            // If it is not in the list of current owners, just remove it
+                            await graphClient.Groups[targetGroup.Id].Owners[owner.Id].Reference.Request().DeleteAsync();
+                        }
+                        catch (ServiceException ex)
+                        {
+                            if (ex.Error.Code == "Request_BadRequest")
+                            {
+                                // Skip any failing removal
+                            }
+                            else
+                            {
+                                throw ex;
+                            }
+                        }
+                    }
+                }
+
+                if (fullListOfOwners.NextPageRequest != null)
+                {
+                    fullListOfOwners = await fullListOfOwners.NextPageRequest.GetAsync();
+                }
+                else
+                {
+                    pageExists = false;
+                }
+            }
+        }
+
+
+        private static List<User> GetUsers(GraphServiceClient graphClient, string[] arrUsers)
+        {
+            if (arrUsers == null || arrUsers.Length == 0)
+            {
+                return new List<User>();
+            }
+            var result = Task.Run(async () =>
+            {
+                var usersResult = new List<User>();
+                var users = await graphClient.Users.Request().GetAsync();
+                while (users.Count > 0)
+                {
+                    foreach (var u in users)
+                    {
+                        if (arrUsers.Any(uc => string.Compare(u.UserPrincipalName, uc, true) == 0))
+                        {
+                            usersResult.Add(u);
+                        }
+                    }
+
+                    if (users.NextPageRequest != null)
+                    {
+                        users = await users.NextPageRequest.GetAsync();
+                    }
+                    else
+                    {
+                        break;
+                    }
+                }
+
+                return usersResult;
+            }).GetAwaiter().GetResult();
+            return result;
+        }
+
 
         static string GetDisplayName(CreateGroupRequest request)
         {
@@ -227,6 +463,9 @@ namespace Cumulus.Monads.Graph
 
             [Display(Description = "AllowToAddGuests")]
             public bool AllowToAddGuests { get; set; }
+
+            public string[] owners { get; set; }
+            public string[] members { get; set; }
         }
 
         public class CreateGroupResponse
@@ -240,5 +479,7 @@ namespace Cumulus.Monads.Graph
             [Display(Description = "Mail of the Office 365 Group")]
             public string Mail { get; set; }
         }
+
+
     }
 }
